problem 265 & some bug fixes

+++ problems/problem3.h
@@ -7,10 +7,10 @@
 
 #include <iostream>
 
-long largest_prime_factor(long n) {
+long long largest_prime_factor(long long n) {
     int i = 2;
-    long lpf = 1;
-    long n_orig = n;
+    long long lpf = 1;
+    long long n_orig = n;
     while (i <= n_orig && n > 1) {
         if (n % i == 0) {
             n /= i;
@@ -23,8 +23,8 @@ long largest_prime_factor(long n) {
 }
 
 void problem_3_solution(bool log) {
-    long n = 600851475143;
-    long a = largest_prime_factor(n);
+    long long n = 600851475143;
+    long long a = largest_prime_factor(n);
     if (log) std::cout << a << std::endl;
 }
 #endif //PROJECT_EULER_PROBLEM3_H
